+++ modules/aws-emr-config/main.tf
@@ -13,7 +13,7 @@ locals {
 # JSON configuration to S3
 resource "aws_s3_bucket_object" "upload_json_config" {
   bucket                 = var.bucket_name_for_root_directory
-  key                    = "config.json"
+  key                    = var.json_configuration_bucket_key
   content                = local.json_config
   content_type           = "application/json"
   server_side_encryption = "AES256"
@@ -23,7 +23,7 @@ resource "aws_s3_bucket_object" "upload_json_config" {
 resource "aws_s3_bucket_object" "upload_hbase_config_script" {
   count  = var.create_static_cluster ? 1 : 0
   bucket = var.bucket_name_for_root_directory
-  key    = "util/upload_hbase_config.sh"
+  key    = var.utility_script_bucket_key
   content = templatefile(
     "${path.module}/upload_hbase_config.sh",
     {