+++ variables.tf
@@ -253,3 +253,15 @@ variable "enable_http_port" {
   description = "EMR services like Ganglia run on the http port"
   default     = false
 }
+
+variable "json_configuration_bucket_key" {
+  type        = string
+  description = "Key (i.e. path) of JSON configuration bucket object in the root directory bucket"
+  default     = "config.json"
+}
+
+variable "utility_script_bucket_key" {
+  type        = string
+  description = "Key (i.e. path) to upload the utility script to"
+  default     = "util/upload_hbase_config.sh"
+}